Revert "Dev memory"
