+++ package.json
@@ -7,7 +7,6 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/axios": "^0.14.0",
     "@types/jest": "^27.5.0",
     "@types/node": "^16.11.33",
     "@types/react": "^18.0.9",
@@ -15,7 +14,6 @@
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.25",
     "@types/uuid": "^8.3.4",
-    "axios": "^0.27.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-icons": "^4.3.1",
