+++ src/App.tsx
@@ -3,13 +3,11 @@ import { Routes, Route, useNavigate } from 'react-router-dom';
 import styled, { createGlobalStyle } from 'styled-components';
 import reset from 'styled-reset';
 
+import TodoList from 'components/TodoList';
+import MenuBar from 'components/MenuBar';
+import Memory from 'components/Memory';
 import { useDispatch, useSelector } from 'react-redux';
 import { todosActions } from 'modules/todos';
-
-import MenuBar from 'components/MenuBar';
-import TodoList from 'components/Todo/TodoList';
-import Memory from 'components/Memory/Memory';
-import { notesActions } from 'modules/memory';
 
 const GlobalStyles = createGlobalStyle`
     ${reset}
@@ -64,13 +62,6 @@ interface ITodo {
   content: string;
 }
 
-interface INote {
-  id: string;
-  title: string;
-  tags: string[];
-  content: string;
-}
-
 function App(): JSX.Element {
   const nav = useNavigate();
   const [todoList, setTodoList] = useState<ITodo[]>([]);
@@ -83,19 +74,11 @@ function App(): JSX.Element {
     (state: { todosSlice: { todos: ITodo[] } }) => state.todosSlice.todos
   );
 
-  const notesSlice = useSelector(
-    (state: { notesSlice: { notes: INote[] } }) => state.notesSlice.notes
-  );
-
   const dispatch = useDispatch();
 
   useEffect(() => {
-    const todosData: string | null = window.localStorage.getItem('todosData');
-    const notesData: string | null = window.localStorage.getItem('notesData');
-
-    if (todosData) dispatch(todosActions.getLocalStorage());
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    if (notesData) dispatch(notesActions.getLocalStorage());
+    const data: string | null = window.localStorage.getItem('todosData');
+    if (data) dispatch(todosActions.getLocalStorage());
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
   // 페이지를 불러 올 때 데이터가 있는 경우 todoList 의 값을 해당 JSON 을 객체로 파싱한 값으로 갱신한다.
@@ -103,10 +86,6 @@ function App(): JSX.Element {
   useEffect(() => {
     window.localStorage.setItem('todosData', JSON.stringify(todoSlice));
   }, [todoSlice]);
-
-  useEffect(() => {
-    window.localStorage.setItem('notesData', JSON.stringify(notesSlice));
-  }, [notesSlice]);
   // 상태가 변경될 때 해당 내용을 로컬 스토리지에 저장한다.
   // 위 두 Effect Hook 으로 인해 서버없이도 브라우저에 저장된다.
